polymer 3 post conversion fixes
