+++ .eslintignore
@@ -1,3 +1,5 @@
 
 test/acceptance/*
 reports
+d2l_lang_plugin/**/*
+langs/**/*
