--- d2l-html-editor-client-framed.html
@@ -1,1 +1,0 @@
-<script type="text/javascript" src="https://s.brightspace.com/lib/ifrau/0.22.1/ifrau/client.js"></script>
