--- d2l-html-editor-client.html
@@ -1,1 +1,0 @@
-<script type="text/javascript" src="./d2l-html-editor-client.js"></script>
