+++ demo/index.html
@@ -14,9 +14,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
 	<meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes">
 	<title>d2l-html-editor Demo</title>
 	<script src="../../@webcomponents/webcomponentsjs/webcomponents-bundle.js"></script>
-	<script type="module" src="./d2l-html-editor-client.js"></script>
 	<script type="module" src="./d2l-editor-wrapper.js"></script>
-	<script type="module" src="../d2l-html-editor.js"></script>
 
 	<custom-style>
 		  <style is="custom-style">
@@ -36,47 +34,12 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
 	  </custom-style>
   </head>
   <body>
+	<p>Demo 1</p>
 	<d2l-editor-wrapper prop1="demo1"></d2l-editor-wrapper>
+
+	<p>Demo 2</p>
+	<d2l-editor-wrapper prop1="demo2"></d2l-editor-wrapper>
 	<p>
-	  Liverpool will take on German giants Bayern Munich in the Champions League last 16, while Manchester United have been drawn against Paris St-Germain.
-
-	  Premier League champions Manchester City and Tottenham also face Bundesliga sides, with City meeting Schalke and Spurs playing Borussia Dortmund.
-
-	  The first legs will take place on 12-13 and 19-20 February, with the return fixtures on 5-6 and 12-13 March.
-
-	  Group winners will be away from home in the first legs.
-
-	  Teams could not be drawn against sides from the same country in the last 16, nor against teams who they played in the group stage.
-
-	  Elsewhere, holders Real Madrid will face Dutch side Ajax, who have reached the knockout stages for the first time since 2005-06, while Serie A champions Juventus face Europa League winners Atletico Madrid.
-
-	  Spanish champions Barcelona meet French side Lyon, while Italian club Roma face Portuguese champions Porto.
-	  Liverpool will take on German giants Bayern Munich in the Champions League last 16, while Manchester United have been drawn against Paris St-Germain.
-
-	  Premier League champions Manchester City and Tottenham also face Bundesliga sides, with City meeting Schalke and Spurs playing Borussia Dortmund.
-
-	  The first legs will take place on 12-13 and 19-20 February, with the return fixtures on 5-6 and 12-13 March.
-
-	  Group winners will be away from home in the first legs.
-
-	  Teams could not be drawn against sides from the same country in the last 16, nor against teams who they played in the group stage.
-
-	  Elsewhere, holders Real Madrid will face Dutch side Ajax, who have reached the knockout stages for the first time since 2005-06, while Serie A champions Juventus face Europa League winners Atletico Madrid.
-
-	  Spanish champions Barcelona meet French side Lyon, while Italian club Roma face Portuguese champions Porto.
-	  Liverpool will take on German giants Bayern Munich in the Champions League last 16, while Manchester United have been drawn against Paris St-Germain.
-
-	  Premier League champions Manchester City and Tottenham also face Bundesliga sides, with City meeting Schalke and Spurs playing Borussia Dortmund.
-
-	  The first legs will take place on 12-13 and 19-20 February, with the return fixtures on 5-6 and 12-13 March.
-
-	  Group winners will be away from home in the first legs.
-
-	  Teams could not be drawn against sides from the same country in the last 16, nor against teams who they played in the group stage.
-
-	  Elsewhere, holders Real Madrid will face Dutch side Ajax, who have reached the knockout stages for the first time since 2005-06, while Serie A champions Juventus face Europa League winners Atletico Madrid.
-
-	  Spanish champions Barcelona meet French side Lyon, while Italian club Roma face Portuguese champions Porto.
 	  Liverpool will take on German giants Bayern Munich in the Champions League last 16, while Manchester United have been drawn against Paris St-Germain.
 
 	  Premier League champions Manchester City and Tottenham also face Bundesliga sides, with City meeting Schalke and Spurs playing Borussia Dortmund.
