+++ package.json
@@ -21,8 +21,7 @@
     "serve": "polymer serve",
     "lint": "npm run lint:wc && npm run lint:js",
     "lint:js": "eslint . --ext .js,.html test/**/*.js test/**/*.html demo/**/*.js demo/**/*.html",
-    "lint:js": "eslint . --ext .js,.html test/**/*.js test/**/*.html demo/**/*.js demo/**/*.html",
-    "lint:wc": "polymer lint",
+    "lint:wc": "polymer lint *.js",
     "test": "npm run lint && npm run test:polymer:local",
     "test:polymer:local": "cross-env LAUNCHPAD_BROWSERS=chrome polymer test --skip-plugin sauce",
     "test:polymer:sauce": "polymer test --skip-plugin local"
@@ -52,7 +51,8 @@
   },
   "dependencies": {
     "@polymer/polymer": "^3.0.0",
-    "d2l-icons": "BrightspaceUI/icons#semver:^6",
-    "d2l-fetch": "Brightspace/d2l-fetch.git#semver:^2"
+    "d2l-fetch": "Brightspace/d2l-fetch.git#semver:^2",
+    "d2l-fetch-auth": "github:Brightspace/d2l-fetch-auth",
+    "d2l-icons": "BrightspaceUI/icons#semver:^6"
   }
 }
