--- tinymce-import.js
@@ -1,10 +1,0 @@
-/*<script type="text/javascript" src="https://cdn.tinymce.com/4/tinymce.js"></script>*/
-/* <script type="text/javascript" src="http://localhost:3000/js/tinymce/tinymce.js"></script> */
-/*<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/tinymce/4.3.10/tinymce.min.js"></script> */
-/* <script type="text/javascript" src="https://s.brightspace.com/lib/tinymce/4.7.11-a11ychecker.1.2.1-53-powerpaste.3.3.0-283-polymer2-hack/tinymce.min.js"></script> */
-/*
-  FIXME(polymer-modulizer): the above comments were extracted
-  from HTML and may be out of place here. Review them and
-  then delete this comment!
-*/
-;
